Move renderValue to separate module
